+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.24.0
+ - feat: Adding `NesIcons.heart`
+
+
 # 0.23.0
  - fix: `NesDialog` would rely on the "root" theme, which would break if it was not a nes ui one.
  - feat: Dialogs now can be "confirmed" with the enter key, and the cancel (esc key) can be intercepted by the developer now